moved functions from main to module compile
